refactor: modularize MCP tool creation steps
